requirements.txt
asgiref==3.2.3
astroid==2.3.3
cycler==0.10.0
Django==3.0.2
gunicorn==20.0.4
isort==4.3.21
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.0
Pillow==7.0.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
